+++ CHANGELOG.md
@@ -7,5 +7,9 @@
   - Queue
   - Stack
   - Priority Queue
+
 - Command line interface ready for basic server commands as well helper commands for ARgorithm Creation
+  - Programmer actions supported
+  - Server admin actions added
+
 - Documentation Generated from doc strings and hosted